[DEV] Add makefile to run common dev commands
It's a bit annoying to type `ninja -C python/build/cmake-.../` to run
build commands. This adds a top-level makefile that just forwards
commands down to the ninja build system so you can just type `make`.

Building and running the tests from a fresh install should now be as
simple as `make dev-install && make test`.

I also add `make golden-samples` to regenerate the golden sample tests,
which should make it fairly painless to accept a failing sample.
